Cliente Carrito Integration

+++ src/main/java/org/example/proyecto_productos/Clientes/repository/ClienteRepository.java
@@ -4,5 +4,7 @@ package org.example.proyecto_productos.Clientes.repository;
 import org.example.proyecto_productos.Clientes.model.Cliente;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClienteRepository extends CrudRepository<Cliente, Integer> {
+import java.util.Optional;
+
+public interface ClienteRepository extends CrudRepository<Cliente, Long> {
 }
